fix lints
